requirements.txt
Flask>=3.0.0
flask-openapi3>=3.1.1
python-dotenv>=1.0.1